Move services to a services/ directory
