+++ packages/hub/index.ts
@@ -1,3 +1,3 @@
 export { Factory, Container } from './di/container';
-export { Session } from './session';
+export { Session } from './services/session';
 export { CardstackError as Error } from './error';
