+++ packages/hub/services/shutdown-helper.ts
